Refactor package structure: move classes to component subpackages for better organization and clarity
